Clean up dependencies.
While working on the dependency diagram in docs I decided that
three.js and flux-three-plugins need to be bundled so that they
can't get out of sync due to user error. I also decided to move
geometry builder from flux-viewport to flux-json-to-three.
https://docs.google.com/document/d/
1qf6PGJN54buSGFcrcRCZnruY4yTAsNRzoMtEpyE2VFY/edit#
By moving it we can remove the dependency of flux-viewport
directly on modelingjs. Also the geometry builder makes
more sense as part of json-to-three and additionally
reduces the interface of that modules.

Change-Id: I69f16e94c90d6ed8096510f3cbb4039db88c9834