LIB3D-819: Revit Element display in viewport.
Updating package version to pickup latest version of flux-json-to-three.

:maker-facing:

QA: Make sure that a single revit elment that contains a mesh should automatically
activate the 3D view in the toolbar. When switching to that view the elements
mesh should be displayed. When data inspector contains an array of revit
elements, any and all elements that contain a mesh should be displayed in the
viewport. Display mesh is located at object.geometryParameters.geometry
in the revit element.

Change-Id: Ie34517e0afaebba5691ea3bc0685d4c22a2bbb03