Improve scene object structure - LIB3D-1017
npm version 0.27.0
Scene creation was a recursive process that involved a lot of merging of
Object3D, which was hard to maintain. I have re-written it to be a two pass
algorithm that first creates all the scene objects and then links them up
acording to the scene graph. This also means the final Object3D tree
better matches the actual scene data which makes applications like
selection much easier to implement.
Update documentation.
Update flux-three-plugins for camera fix.

:not-maker-facing:

QA:
 1. None.
Change-Id: I119117ea95f7d1ccb322a2543811c6282c9d3d3f