Update three.js - LIB3D-905
npm version 0.28.0
Update to three.js r82 which contains es6 modules. And fix for node 6.9.1
Eventually this will allow tree shaking, but much of the optimization
is currently diminished due to limitations in rollup that can not
handle certain patterns currently used in three.js. Also we are
still treating three as an external global.

:not-maker-facing:

QA:
 1. Create the 3D QA Module app, it should render as usual.

Change-Id: Ifdef932fe17342a5e1100b83d6b2e4cd6e5164bf