Skip entity warnings on invalid scenes - LIB3D-1089
npm version 0.33.2
When JSON data is a scene, but the scene is invalid, it is rendered
as entities. There tend to be errors when partially valid scenes have
elements that are then invalid entities. Thus the new strategy in this
case is to ignore the entity errors and present the scene errors to the
user, since those are most helpful.
Also, pick up an update to the validator to have better messages.

:hp: :viewport:

Code Review:
1. Check out related CLs.
2. make run-viewport
QA:
1. Render the liberty.json from JIRA.
2. The text of the error in the viewport should contain one message,
   not two. (like streamlined-errors.png, not original-errors.png)

Change-Id: Icc96b8e6089b65ca800688c3d344dcbabd01a6bc